test(e2e): give the draft a model before expecting it to publish
`agents.spec.ts:573` failed on the assertion this branch added: the
publish dialog never opened, because `/validate` refused the draft with
`No model selected - pick one before publishing`. The MCP binding this
test is about was never the problem - the seeded `E2E Draft` is created
with a name and nothing else, so it has no model profile until
`models.spec.ts` runs, which is alphabetically later. It now names the
seeded profile and waits for the autosave before clicking Publish.

The refusal was invisible until this branch because the check for it
asserted nothing: `toBeHidden()` passes on an element that has not
rendered yet, and at click time the problems panel has not. The failure
screenshot from CI shows that panel listing the missing model *while*
that line passed. Replaced with an assertion on the `/validate` answer,
whose message carries the refusal body - a spec that cannot publish now
says which problem stopped it instead of timing out for 15 seconds on a
dialog that was never going to open.

`selectSavedModel` and `savedModels` move from `models.spec.ts` into
`helpers.ts` unchanged, because a model is what any spec needs before it
can publish anything at all.

Verified: `make lint-frontend` green. The e2e answer comes from CI - it
needs a seeded backend, which is why it is not in `make check`.

Part-of #519
